requirements.txt
distro==1.9.0
lxml==5.4.0
Nuitka==2.7
ordered-set==4.1.0
packaging==25.0
pillow==11.2.1
scikit-build==0.18.1
setuptools==80.3.1
sv-ttk==2.6.0
tkfontawesome==0.3.2
tksvg==0.7.4
wheel==0.45.1
zstandard==0.23.0
